src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/custom/isolated-world-expected.txt

CONSOLE ERROR: line 53: Uncaught TypeError: document.registerElement is not a function
Tests that accessing custom elements from an isolated world does not cause worlds to collide and destroy the galaxy.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.testRunner is non-null.
FAIL result should be 9 (of type number). Was undefined (of type undefined).
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
